+++ Campaign/extensions/campaign-module/src/Banner.php
@@ -310,6 +310,16 @@ class Banner extends Model implements Searchable
         if ($this->js) {
             $snippetsToProcess = array_merge($snippetsToProcess, $this->extractSnippetCodes($this->js));
         }
+        if ($this->js_includes) {
+            foreach ($this->js_includes as $jsInclude) {
+                $snippetsToProcess = array_merge($snippetsToProcess, $this->extractSnippetCodes($jsInclude));
+            }
+        }
+        if ($this->css_includes) {
+            foreach ($this->css_includes as $cssInclude) {
+                $snippetsToProcess = array_merge($snippetsToProcess, $this->extractSnippetCodes($cssInclude));
+            }
+        }
 
         // Extract snippets from template text and CSS fields
         $snippetsToProcess = array_merge($snippetsToProcess, $this->extractSnippetsFromTemplate());