+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### [Campaign]
 
 - Fixed broken use of snippets within snippets, introduced in `4.2.0`. remp/remp#1441
+- Added support for use of snippets in JS includes. remp/remp#1441
 
 ### [Mailer]